+++ CHANGES.md
@@ -2,6 +2,7 @@
 * remove Freemius from the autoloader
 * ensure `is_plugin_active()` is available when needed
 * utilize new `class Ignore` from Git Updater
+* uses new `class Fragen\Git_Updater\Shim` for PHP 5.6 compatibility, will remove when WP core changes minimum requirement
 
 #### 1.1.1 / 2021-06-14
 * utilize new `class Ignore` in Git Updater